Few modification and changing for size_t to uint.

+++ math/dataset.cc
@@ -11,19 +11,18 @@
 #include <gsl/gsl_combination.h>
 
 #include "dataset.hh"
-#include "pputils.hh"
 
 using std::cout;
 using std::cerr;
 using std::ofstream;
 
-MSZDataSet::MSZDataSet(double **m, size_t nrows, size_t ncols, size_t outputs) 
+MSZDataSet::MSZDataSet(double **m, uint nrows, uint ncols, uint outputs) 
   : nrows(nrows), rfeatures(ncols - outputs), ncols(ncols), outputs(outputs),
     stdDone(false), oStdDone(false) {
 
   // Recomputing matrix to be made of columns instead of lines
   matrix = new double* [ncols];
-  for(size_t c = 0; c < ncols; c++)
+  for(uint c = 0; c < ncols; c++)
     setMColumn(c, new double [nrows]);
   
   // Copy the elements
@@ -43,7 +42,7 @@ MSZDataSet::MSZDataSet(double **m, size_t nrows, size_t ncols, size_t outputs)
   printRawMatrix();
 
   // Delete old matrix
-  for(size_t r = 0; r < nrows; r++)
+  for(uint r = 0; r < nrows; r++)
     delete[] m[r];
   delete[] m;
 }
@@ -55,11 +54,11 @@ MSZDataSet::MSZDataSet(const MSZDataSet& ds)
 
   // Now we just need to copy the matrix.
   matrix = new double* [ncols];
-  for(size_t c = 0; c < ncols; c++)
+  for(uint c = 0; c < ncols; c++)
     setMColumn(c, new double [nrows]);
 
-  for(size_t r = 0; r < nrows; r++)
-    for(size_t c = 0; c < ncols; c++)
+  for(uint r = 0; r < nrows; r++)
+    for(uint c = 0; c < ncols; c++)
       setMValue(r, c, ds.getMValue(r, c));
 
 }
@@ -69,7 +68,7 @@ MSZDataSet::~MSZDataSet() {
   // Please God, do not let this Seg Fault!
 
   // Clean up matrix by deleting each line.
-  for(size_t c = 0; c < ncols; c++) 
+  for(uint c = 0; c < ncols; c++) 
     delete[] getMColumn(c);
 	 
   delete[] matrix;
@@ -79,9 +78,9 @@ MSZDataSet::~MSZDataSet() {
 void MSZDataSet::printRawMatrix() {
 
   cerr << "Printing all matrix [" << nrows << ", " << ncols << "]\n";
-  for(size_t r = 0; r < nrows; r++) {
+  for(uint r = 0; r < nrows; r++) {
     cerr << "[Row " << r << "] ";
-    for(size_t c = 0; c < ncols; c++) {
+    for(uint c = 0; c < ncols; c++) {
       cerr << getMValue(r, c) << " ";
     }
     cerr << "\n";
@@ -98,8 +97,8 @@ void MSZDataSet::dumpPlotFiles(const vector<string> &labels, const string &prefi
   }
 
   // Dump the files for plotting features against output
-  for(size_t f1 = 0; f1 < outputs; f1++) {
-    for(size_t f2 = (f1 < outputs ? outputs : f1+1); f2 < ncols; f2++) {
+  for(uint f1 = 0; f1 < outputs; f1++) {
+    for(uint f2 = (f1 < outputs ? outputs : f1+1); f2 < ncols; f2++) {
       ofstream file;
       file.open((prefix + "@" + labels[f1] + "@" + labels[f2] + ".dat").c_str());
       
@@ -114,7 +113,7 @@ void MSZDataSet::dumpPlotFiles(const vector<string> &labels, const string &prefi
 	   << "# " << labels[f2] << "\t\t" << labels[f1] << "\n";
       
       // For each instance
-      for(size_t i = 0; i < nrows; i++) 
+      for(uint i = 0; i < nrows; i++) 
 	file << getMValue(i, f2) << "\t\t" << getMValue(i, f1) << "\n";
 
       file.close();
@@ -122,18 +121,18 @@ void MSZDataSet::dumpPlotFiles(const vector<string> &labels, const string &prefi
   }
 }
 
-void MSZDataSet::printSolverStats(size_t timeout, const vector<string> &slabels) {
+void MSZDataSet::printSolverStats(uint timeout, const vector<string> &slabels) {
 
   if(slabels.size() != outputs) {
     cerr << "printSolverStats: You're passing in " << slabels.size() << " solver labels but there are " << outputs << " outputs in dataset.\n";
     return;
   }
 
-  vector<size_t> nbTimeouts(outputs, 0);
-  vector<size_t> nbOtherErrors(outputs, 0);
+  vector<uint> nbTimeouts(outputs, 0);
+  vector<uint> nbOtherErrors(outputs, 0);
 
-  for(size_t s = 0; s < outputs; s++) {
-    for(size_t i = 0; i < nrows; i++) {
+  for(uint s = 0; s < outputs; s++) {
+    for(uint i = 0; i < nrows; i++) {
       if(getMValue(i, s) == timeout)
 	nbTimeouts[s]++;
       else if(getMValue(i, s) == 2*timeout)
@@ -141,7 +140,7 @@ void MSZDataSet::printSolverStats(size_t timeout, const vector<string> &slabels)
     }
   }
 
-  for(size_t s = 0; s < outputs; s++) {
+  for(uint s = 0; s < outputs; s++) {
     cout << "Solver: " << slabels[s] << "\n"
 	 << "\t\tTimeouts: " << nbTimeouts[s] << " (" << (((double)(nbTimeouts[s])) / nrows)*100.0 << ")\n"
 	 << "\t\tOther Errors: " << nbOtherErrors[s] << " (" << (((double)(nbOtherErrors[s])) / nrows)*100.0 << ")\n"
@@ -150,24 +149,24 @@ void MSZDataSet::printSolverStats(size_t timeout, const vector<string> &slabels)
 
 }
 
-void MSZDataSet::dumpPlotFiles(char **labels, size_t len, char *prefix) const {
+void MSZDataSet::dumpPlotFiles(char **labels, uint len, char *prefix) const {
 
   vector<string> vec(len);
 
-  for(size_t i = 0; i < len; i++)
+  for(uint i = 0; i < len; i++)
     vec[i] = string(labels[i]);
 
   dumpPlotFiles(vec, string(prefix));
 }
 
-double MSZDataSet::getOutputValue(size_t row, size_t col) const {
+double MSZDataSet::getOutputValue(uint row, uint col) const {
   assert(row < nrows);
   assert(col < outputs);
 
   return getMValue(row, col);
 }
 
-double MSZDataSet::getFeatureValue(size_t row, size_t col) const {
+double MSZDataSet::getFeatureValue(uint row, uint col) const {
   assert(row < nrows);
   assert(col < ncols - outputs);
 
@@ -179,11 +178,11 @@ void MSZDataSet::standardize() {
   cout << "Standardizing features forall k . x_k = `q(x_k - x')/sdv_i(x_i)...\n";
 
   if(!stdDone) {
-    for(size_t c = outputs; c < ncols; c++) {
+    for(uint c = outputs; c < ncols; c++) {
 
       // Compute column mean and compute column variance.
       double mean = 0.0;
-      for(size_t r = 0; r < nrows; r++) 
+      for(uint r = 0; r < nrows; r++) 
 	mean += getMValue(r, c);
       mean /= nrows;
 
@@ -191,14 +190,14 @@ void MSZDataSet::standardize() {
 
       // Compute column standard deviation. 
       double sdv = 0.0;
-      for(size_t r = 0; r < nrows; r++) 
+      for(uint r = 0; r < nrows; r++) 
 	sdv += gsl_pow_2(getMValue(r, c) - mean);
       sdv /= nrows;
       sdv = sqrt(sdv);
 
       cout << "Scaling = " << sdv << "\n";
 
-      for(size_t r = 0; r < nrows; r++) 
+      for(uint r = 0; r < nrows; r++) 
 	setMValue(r, c, (getMValue(r, c) - mean) / sdv);
 	
     }
@@ -210,7 +209,7 @@ void MSZDataSet::standardize() {
 
 }
 
-void MSZDataSet::removeFeatures(const vector<size_t> &keepVec) {
+void MSZDataSet::removeFeatures(const vector<uint> &keepVec) {
   // We need to remove the feature indexes in vec from the current data.
   
   cout << "Keeping features in dataset: ";
@@ -220,34 +219,34 @@ void MSZDataSet::removeFeatures(const vector<size_t> &keepVec) {
   // - incrementing all indices by output
   // - sorting
   // - adding output indices to the beginning of vector
-  vector<size_t> vec(keepVec);
-  for(vector<size_t>::iterator it = vec.begin(); 
+  vector<uint> vec(keepVec);
+  for(vector<uint>::iterator it = vec.begin(); 
       it != vec.end();
       it++)
     *it += outputs;
   sort(vec.begin(), vec.end());        // sorts 
   vec.insert(vec.begin(), outputs, 0); // adds positions
-  for(size_t o = 0; o < outputs; o++)   
+  for(uint o = 0; o < outputs; o++)   
     vec[o] += o;                       // sets positions
 
   if(vec.size() == ncols) { 
     cout << "ALL\n";
     return;
   } else {
-    copy(vec.begin(), vec.end(), std::ostream_iterator<size_t>(cout, " "));
+    copy(vec.begin(), vec.end(), std::ostream_iterator<uint>(cout, " "));
     cout << std::endl;
   }
 
   // Count the number of raw features about to be removed
-  size_t rawToRemove = 0;
-  for(vector<size_t>::const_iterator it = vec.begin(); 
+  uint rawToRemove = 0;
+  for(vector<uint>::const_iterator it = vec.begin(); 
       it != vec.end();
       it++)
     if(*it >= outputs && *it < outputs + rfeatures) 
       rawToRemove++;
 
   rfeatures -= rawToRemove;
-  const size_t oldNCols = ncols;
+  const uint oldNCols = ncols;
   ncols = vec.size();
   
   // Now, we just need to delete the columns referenced in the vector
@@ -257,12 +256,12 @@ void MSZDataSet::removeFeatures(const vector<size_t> &keepVec) {
   // in the new matrix and the vector contents the origin column on the
   // old matrix. This means that we need to copy the column v[i] of old matrix
   // to column i of the new matrix.
-  for(size_t c = 0; c < ncols; c++) {
+  for(uint c = 0; c < ncols; c++) {
     newMatrix[c] = getMColumn(vec[c]);
     setMColumn(vec[c], 0); // Those which are not 0 are the columns 
                         // that should be deleted afterwards
   }
-  for(size_t c = 0; c < oldNCols; c++)
+  for(uint c = 0; c < oldNCols; c++)
     if(getMColumn(c) != 0) delete[] getMColumn(c);
   
   delete[] matrix;
@@ -274,8 +273,8 @@ void MSZDataSet::standardizeOutputs() {
   cout << "Standardizing outputs... ";
 
   if(!stdDone) {
-    for(size_t c = 0; c < outputs; c++) {
-      for(size_t r = 0; r < nrows; r++) {
+    for(uint c = 0; c < outputs; c++) {
+      for(uint r = 0; r < nrows; r++) {
 	setMValue(r, c, log(getMValue(r, c)));
       }
     }
@@ -287,20 +286,20 @@ void MSZDataSet::standardizeOutputs() {
 
 }
 
-void MSZDataSet::expand(size_t n) { 
-  vector<size_t> pvec(ncols - outputs); 
-  for(size_t i = 0; i < pvec.size(); i++)
+void MSZDataSet::expand(uint n) { 
+  vector<uint> pvec(ncols - outputs); 
+  for(uint i = 0; i < pvec.size(); i++)
     pvec[i] = i;
 
   expandOnPartition(n, pvec);
 }
 
-void MSZDataSet::expand(size_t n, const vector<vector<size_t> > &pvec) {
-  for(size_t i = 0; i < pvec.size(); i++)
+void MSZDataSet::expand(uint n, const vector<vector<uint> > &pvec) {
+  for(uint i = 0; i < pvec.size(); i++)
     expandOnPartition(n, pvec[i]);
 }
 
-void MSZDataSet::expandOnPartition(size_t k, const vector<size_t> &pvec) {
+void MSZDataSet::expandOnPartition(uint k, const vector<uint> &pvec) {
   
   // expansion is quadratic at the minimum and 
   // and be bigger than partition size.
@@ -311,18 +310,18 @@ void MSZDataSet::expandOnPartition(size_t k, const vector<size_t> &pvec) {
 
   // Output information
   cout << "Expanding feature set by order " << k << " on partition:\n";
-  copy(pvec.begin(), pvec.end(), std::ostream_iterator<size_t>(cout, " "));
+  copy(pvec.begin(), pvec.end(), std::ostream_iterator<uint>(cout, " "));
   cout << "\n";
 
   // Reallocate the data.
   // If we have N features, then, to expand to a order-K polynomial
   // we are adding 
   // n + (n !) / (k! (n-k!))
-  size_t num = 1;
-  for(size_t i = 0; i < k; i++) num *= rfeatures - i; // for big k hell breaks loose
-  size_t den = 1;
-  for(size_t i = k; i > 0; i--) den *= i; 
-  size_t nNewF = rfeatures;
+  uint num = 1;
+  for(uint i = 0; i < k; i++) num *= rfeatures - i; // for big k hell breaks loose
+  uint den = 1;
+  for(uint i = k; i > 0; i--) den *= i; 
+  uint nNewF = rfeatures;
   nNewF += num / den;
   ncols += nNewF; // Update number of columns
 
@@ -330,32 +329,32 @@ void MSZDataSet::expandOnPartition(size_t k, const vector<size_t> &pvec) {
   
   // Reallocating main vector and copying all the others to their initial places.
   double **newMatrix = new double* [ncols];
-  for(size_t c = 0; c < ncols - nNewF; c++)
+  for(uint c = 0; c < ncols - nNewF; c++)
     newMatrix[c] = getMColumn(c);
   delete[] matrix;
   matrix = newMatrix;
 
-  for(size_t nc = ncols - nNewF; nc < ncols; nc++)
+  for(uint nc = ncols - nNewF; nc < ncols; nc++)
     setMColumn(nc, new double [nrows]);
 
   // Compute the cross products
   // To do this we compute all the indices combinations and use them
   // to compute the cross product on columns defined in pvec.
-  size_t currColumn = rfeatures + outputs;
-  const size_t n = pvec.size();
+  uint currColumn = rfeatures + outputs;
+  const uint n = pvec.size();
   gsl_combination * c;
   c = gsl_combination_calloc (n, k);
-  size_t *ind = new size_t [k];
+  uint *ind = new uint [k];
   
   // Compute cross product for initial configuration
   // Now using the very nice combination structure from GSL.
   // Up to rev. 121, we used custom combination generator.
   do {
-    for(size_t i = 0; i < k; i++)
+    for(uint i = 0; i < k; i++)
       ind[i] = gsl_combination_get(c, i);
 
     // Compute cross product in current configuration
-    for(size_t r = 0; r < nrows; r++) {
+    for(uint r = 0; r < nrows; r++) {
       const double cprod = computeCrossProduct(r, ind, k, pvec);
       setMValue(r, currColumn, cprod);
     }
@@ -365,8 +364,8 @@ void MSZDataSet::expandOnPartition(size_t k, const vector<size_t> &pvec) {
   delete[] ind;
 
   // Compute Powers
-  for(size_t i = 0; i < pvec.size(); i++) {
-    for(size_t r = 0; r < nrows; r++) {
+  for(uint i = 0; i < pvec.size(); i++) {
+    for(uint r = 0; r < nrows; r++) {
       // Compute powers for pvec[i] feature
       setMValue(r, currColumn, gsl_pow_int(getMValue(r, pvec[i]), k));
     }
@@ -374,14 +373,14 @@ void MSZDataSet::expandOnPartition(size_t k, const vector<size_t> &pvec) {
 
 }
 
-double MSZDataSet::computeCrossProduct(size_t r, size_t *ind, size_t k, const vector<size_t> &vec) {
+double MSZDataSet::computeCrossProduct(uint r, uint *ind, uint k, const vector<uint> &vec) {
   double cp = 1.0; // Cross-product
-  for(size_t i = 0; i < k; i++)
+  for(uint i = 0; i < k; i++)
     cp *= getMValue(r, vec[ind[i]]);
   return cp;
 }
 
-void MSZDataSet::removeTimeouts(size_t timeout, size_t out) {
+void MSZDataSet::removeTimeouts(uint timeout, uint out) {
   assert(out < outputs);
 
   if(oStdDone) {
@@ -390,8 +389,8 @@ void MSZDataSet::removeTimeouts(size_t timeout, size_t out) {
   }
 
   // Count and keep rows without timeouts
-  vector<size_t> keepRows;
-  for(size_t r = 0; r < nrows; r++) {
+  vector<uint> keepRows;
+  for(uint r = 0; r < nrows; r++) {
     if(getMValue(r, out) != timeout) 
       keepRows.push_back(r);
   }
@@ -406,14 +405,14 @@ void MSZDataSet::removeTimeouts(size_t timeout, size_t out) {
 
   // Allocating new matrix
   double **newMatrix = new double* [ncols];
-  for(size_t c = 0; c < ncols; c++)
+  for(uint c = 0; c < ncols; c++)
     newMatrix[c] = new double [keepRows.size()];
 
   nrows = keepRows.size();
 
-  for(size_t c = 0; c < ncols; c++) {
-    size_t nr = 0;
-    for(vector<size_t>::const_iterator rptr = keepRows.begin();
+  for(uint c = 0; c < ncols; c++) {
+    uint nr = 0;
+    for(vector<uint>::const_iterator rptr = keepRows.begin();
 	rptr != keepRows.end();
 	rptr++) {
       assert(*rptr >= nr);
@@ -423,7 +422,7 @@ void MSZDataSet::removeTimeouts(size_t timeout, size_t out) {
   }
 
   // delete old matrix
-  for(size_t c = 0; c < ncols; c++)
+  for(uint c = 0; c < ncols; c++)
     delete [] matrix[c];
   delete [] matrix;
 
@@ -439,7 +438,7 @@ void MSZDataSet::removeTimeouts(size_t timeout, size_t out) {
 /////////////////////////////////////////
 /////////////////////////////////////////
 
-MSZDataSet *createDataSet(double** matrix, size_t nrows, size_t ncols, size_t outputs) {
+MSZDataSet *createDataSet(double** matrix, uint nrows, uint ncols, uint outputs) {
   assert(nrows > 0);
   assert(ncols > 0);
   assert(matrix != 0);
@@ -447,7 +446,7 @@ MSZDataSet *createDataSet(double** matrix, size_t nrows, size_t ncols, size_t ou
   assert(nrows >= ncols - outputs);
   
 #ifndef NDEBUG
-  for(size_t r = 0; r < nrows; r++)
+  for(uint r = 0; r < nrows; r++)
     assert(matrix[r] != 0);
 #endif // NDEBUG
