+++ math/forwardselection.hh
@@ -7,6 +7,7 @@
 #include <gsl/gsl_matrix.h>
 
 #include "dataset.hh"
+#include "pputils.hh"
 
 using std::vector;
 
@@ -17,21 +18,21 @@ using std::vector;
 // dataset with multiple inputs and multiple outputs.
 class ForwardSelection {
 public:
-  ForwardSelection(const MSZDataSet &, size_t);
+  ForwardSelection(const MSZDataSet &, uint);
   ~ForwardSelection();
 
   /// Given a double which is the fin target to add regressors,
   /// returns a vector with the indices of the regressors
   /// which should be included in the model.
-  vector<size_t> run(double);
+  vector<uint> run(double);
   
 private:
-  double computeFtest(const vector<bool> &, double, size_t, double *);
+  double computeFtest(const vector<bool> &, double, uint, double *);
 
   gsl_matrix *fmatrix;///< Feature Matrix
   gsl_vector *ovec;   ///< Output Vector
   double SSt;     ///< Total Corrected Sum of Squares
-  size_t initRegressor;     ///< Initial FS Regressor (highest correlated with output)
+  uint initRegressor;     ///< Initial FS Regressor (highest correlated with output)
   double initSSr;     ///< Initial regression sum of squares 
 };
